+++ src/Configuration/Dotenv/Internal.hs
@@ -1,0 +1,12 @@
+module Configuration.Dotenv.Internal
+  ( module Configuration.Dotenv.Types
+  , module Configuration.Dotenv.Parse
+  , module Configuration.Dotenv.ParsedVariable
+  , module Configuration.Dotenv.Text
+  )
+where
+
+import           Configuration.Dotenv.Parse
+import           Configuration.Dotenv.ParsedVariable
+import           Configuration.Dotenv.Text
+import           Configuration.Dotenv.Types