+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## MASTER
+## Dotenv 0.11.0.1
+### Modified
+* Export internal module `Configuration.Dotenv.Internal` which exports all
+the internal modules.
+* Export `configParser` from `Configuration.Dotenv`.
+
 ## Dotenv 0.11.0.0
 ### Modified (Breaking change - new behavior)
 * Take last rather than first env var in dotenv file (reported by @rudymatela and